Remove torch from core dependencies (fixes #23) (#36)
* Remove torch from core dependencies (issue #23)

- Move torch and tensorboard to optional dependency groups (contrib-rl, contrib-mi)
- Use TYPE_CHECKING to make torch import conditional in stat_tracker.py
- Add runtime check in TensorboardStatTracker with helpful error message
- Reduces core installation size by ~71MB for users who don't need TensorBoard

Fixes #23

* Refactor stat tracking to use optional dependencies

Based on review feedback from @joshgreaves:

- Created ares/experiment_tracking module with cleaner separation:
  - stat_tracker.py: Contains StatTracker interface and base implementations
  - tensorboard.py: Contains TensorboardStatTracker (torch-dependent)
  - __init__.py: Gracefully handles optional tensorboard import

- Changed pyproject.toml to use [project.optional-dependencies]:
  - tensorboard = ["tensorboard>=2.20.0", "torch>=2.9.1"]
  - This is for optional runtime dependencies (vs dependency-groups for dev)

- Removed TYPE_CHECKING pattern since import will fail gracefully
  if torch is not installed

- Updated all imports throughout codebase:
  - Changed from ares.code_agents.stat_tracker
  - To ares.experiment_tracking

- Removed old stat_tracker.py from code_agents

Users can install TensorBoard support with: pip install ares[tensorboard]

* Keep __init__.py empty per repo design decision
