+++ pyproject.toml
@@ -34,11 +34,10 @@ dependencies = [
     "pyyaml>=6.0.3",
     "swebench>=4.1.0",
     "tenacity>=9.1.2",
-    # TODO: Add experiment logging that doesn't require torch.
-    # This is only needed for tensorboard support.
-    "tensorboard>=2.20.0",
-    "torch>=2.9.1",
 ]
+
+[project.optional-dependencies]
+tensorboard = ["tensorboard>=2.20.0", "torch>=2.9.1"]
 
 [project.urls]
 Homepage = "https://github.com/withmartian/ares"
