+++ src/ares/experiment_tracking/stat_tracker.py
@@ -4,13 +4,11 @@ import asyncio
 import collections
 from collections.abc import Generator
 import contextlib
-import dataclasses
 import logging
 import time
 from typing import Protocol
 
 import numpy as np
-from torch.utils import tensorboard
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -51,37 +49,6 @@ class LoggingStatTracker(StatTracker):
                     continue
 
                 _LOGGER.info("%s: %s", k, np.percentile(np.array(v), [0, 25, 50, 75, 100]))
-            self._stats.clear()
-
-    @contextlib.contextmanager
-    def timeit(self, name: str) -> Generator:
-        start_time = time.time()
-        yield
-        end_time = time.time()
-        self._stats[name].append(end_time - start_time)
-
-    def scalar(self, name: str, value: float) -> None:
-        self._stats[name].append(value)
-
-
-@dataclasses.dataclass
-class TensorboardStatTracker(StatTracker):
-    metric_writer: tensorboard.SummaryWriter
-    period_seconds: float = 60
-
-    def __post_init__(self) -> None:
-        self._stats: collections.defaultdict[str, list[float]] = collections.defaultdict(list)
-        self._task = asyncio.create_task(self._track())
-
-    async def _track(self) -> None:
-        step = 0
-        while True:
-            step += 1
-            await asyncio.sleep(self.period_seconds)
-
-            # We don't need to sort the keys for this stat tracker.
-            for k, v in self._stats.items():
-                self.metric_writer.add_histogram(k, np.asarray(v), global_step=step)
             self._stats.clear()
 
     @contextlib.contextmanager
